more AbstractLatentWorker to its own file
This is easier to work with when it has its own file
