+++ master/buildbot/worker/ec2.py
@@ -32,7 +32,7 @@ from twisted.python import log
 
 from buildbot import config
 from buildbot.interfaces import LatentWorkerFailedToSubstantiate
-from buildbot.worker.base import AbstractLatentWorker
+from buildbot.worker import AbstractLatentWorker
 from buildbot.worker_transition import reportDeprecatedWorkerNameUsage
 
 try:
